reorganized the lib, nuget and redist folders by release. All previous releases are available in repo
